requirements.txt
PyQt5>=5.15.0
PyYAML>=6.0
Pillow>=8.0.0
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
pandas>=1.1.0

# YOLOv8训练相关依赖
ultralytics>=8.0.0